Debugging risorse sul subject e upload di TestDopoDebugging

+++ source/Proxy/src/main/java/anaws/Proxy/ProxySubject/ResponseHandler.java
@@ -35,7 +35,7 @@ public class ResponseHandler implements CoapHandler {
 		if(this.registration.getType().equals("battery")) {
 			//Updating battery
 			Log.info("ResponseHandler", "Update battery level, new value: " + response.getResponseText());
-			this.registration.getSensorNode().updateBattery(Double.valueOf(response.getResponseText()), proxyObserver);
+			this.registration.getSensorNode().updateBattery(Integer.valueOf(response.getResponseText()), proxyObserver);
 		}
 		else {
 			String Message = response.getResponseText();
